requirements.txt
discordwebhook
requests
python-dotenv
